requirements.txt
pyyaml==3.10
mysql-python==1.2.5
numpy==1.8.2
pytz==2014.7
pandas==0.14.1
six==1.7.3
python-mimeparse==0.1.4
python-dateutil==2.2
Django==1.6.5
django-tastypie==0.11.1
django-configurations==0.8
